[index][linearization] move to its own module
